+++ cocos/2d/CCDirector.h
@@ -379,7 +379,7 @@ protected:
     void showStats();
     void createStatsLabel();
     void calculateMPF();
-    void getFPSImageData(unsigned char** datapointer, long* length);
+    void getFPSImageData(unsigned char** datapointer, size_t* length);
     
     /** calculates delta time since last time it was called */    
     void calculateDeltaTime();
