+++ cocos/2d/platform/CCFileUtils.h
@@ -98,7 +98,7 @@ public:
      *  @return Upon success, a pointer to the data is returned, otherwise NULL.
      *  @warning Recall: you are responsible for calling free() on any Non-NULL pointer returned.
      */
-    virtual unsigned char* getFileDataFromZip(const char* zipFilePath, const char* filename, long *size);
+    virtual unsigned char* getFileDataFromZip(const char* zipFilePath, const char* filename, size_t *size);
 
     
     /** Returns the fullpath for a given filename.
